requirements.txt
google-generativeai>=0.3.0
imap-tools>=1.5.0
loguru>=0.7.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0